refactor: refactor zonky extractor to use generic tar_xz_extractor
