+++ examples/zonky/src/main.rs
@@ -9,7 +9,7 @@ use postgresql_embedded::{PostgreSQL, Result, Settings};
 async fn main() -> Result<()> {
     let settings = Settings {
         releases_url: zonky::URL.to_string(),
-        version: VersionReq::parse("=16.2.0")?,
+        version: VersionReq::parse("=16.3.0")?,
         ..Default::default()
     };
     let mut postgresql = PostgreSQL::new(settings);
