+++ postgresql_archive/src/extractor/mod.rs
@@ -1,8 +1,10 @@
 mod model;
 pub mod registry;
 mod tar_gz_extractor;
+mod tar_xz_extractor;
 mod zip_extractor;
 
 pub use model::ExtractDirectories;
 pub use tar_gz_extractor::extract as tar_gz_extract;
+pub use tar_xz_extractor::extract as tar_xz_extract;
 pub use zip_extractor::extract as zip_extract;
